pyproject.toml
# SPDX-FileCopyrightText: 2026 Zoe Nickson <zoe.nickson@sidingsmedia.com>
# SPDX-License-Identifier: MIT

[project]
name = "finance-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Zoe Nickson",email = "zoe.nickson@sidingsmedia.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pylint (>=4.0.5,<5.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-env (>=1.6.0,<2.0.0)"
]

[tool.fastapi]
entrypoint = "app.main:create_app"

[tool.poetry]
packages = [
    { include = "app" },
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = "--cov --cov-branch --cov-report term-missing"

[tool.pytest_env]
DEBUG = "true"

[tool.codespell]
skip = 'poetry.lock'
count = true
quiet-level = 3

[tool.pylint]
init-hook='import sys; sys.path.append(".")'
